remove old page context
