+++ src/theme.js
@@ -1,0 +1,18 @@
+import { createMuiTheme } from '@material-ui/core/styles'
+
+// Create a theme instance.
+const theme = createMuiTheme({
+    palette: {
+        // primary: {
+        // },
+        // secondary: {
+        // },
+    },
+    typography: {
+        fontFamily: "'Open Sans', sans-serif",
+        fontFamilyMono:
+            'SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace',
+    },
+})
+
+export default theme
